+++ NEWS.rst
@@ -13,6 +13,8 @@ The format is based on `Keep a Changelog
 
 Added
 -----
+- Welcome dialog for users to direct users to the configuration dialog `#43
+  <https://github.com/orontee/argos/issues/43>`_
 - CSS identifiers to allow for style customization  `#72
   <https://github.com/orontee/argos/issues/72>`_
 - Display labels with links when tracklist is empty `#71